backend/requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
pyjwt==2.8.0
pytz==2023.3
sqlparse==0.4.4
